+++ src/ltbio/ml/datasets/BiosignalDataset.py
@@ -13,12 +13,12 @@
 # Last Updated: 05/08/2022
 
 # ===================================
-from abc import ABC, abstractmethod
+from abc import ABC
 from typing import Sequence
 
 from numpy import ndarray
 from torch import Generator, randperm
-from torch.utils.data.dataset import Dataset, random_split, Subset
+from torch.utils.data.dataset import Dataset
 
 from ltbio.biosignals import Biosignal
