--- src/ltbio/ml/models/__init__.py
@@ -1,3 +1,0 @@
-# Quick shortcuts to classes
-from ltbio.ml.models.SkLearnModel import SkLearnModel
-from ltbio.ml.models.TorchModel import TorchModel
