+++ src/ltbio/ml/supervised/__init__.py
@@ -1,0 +1,3 @@
+# Quick shortcuts to classes
+from ltbio.ml.supervised.SupervisedTrainConditions import SupervisedTrainConditions
+from ltbio.ml.supervised.SupervisingTrainer import SupervisingTrainer
