+++ src/ltbio/ml/supervised/models/__init__.py
@@ -1,0 +1,3 @@
+# Quick shortcuts to classes
+from ltbio.ml.supervised.models.SkLearnModel import SkLearnModel
+from ltbio.ml.supervised.models.TorchModel import TorchModel
