--- src/ltbio/ml/trainers/__init__.py
@@ -1,4 +1,0 @@
-# Quick shortcuts to classes
-from ltbio.ml.trainers.SupervisingTrainer import SupervisingTrainer
-from ltbio.ml.trainers.SupervisedTrainConditions import SupervisedTrainConditions
-#from ltbio.ml.trainers.SupervisedTrainResults import SupervisedTrainResults
